Add sync_notion kwarg

+++ app/syncers/google.py
@@ -20,7 +20,7 @@ class GoogleSyncer:
         self.last_sync = None
         self.synced_tasks = []
 
-    def sync_task(self, g_task: GoogleTask, fix_parent=True) -> GoogleTask:
+    def sync_task(self, g_task: GoogleTask, fix_parent=True, sync_notion=True) -> GoogleTask:
         logger.debug(f'Syncing task "{g_task.title}"')
 
         try:
@@ -33,31 +33,36 @@ class GoogleSyncer:
                 logger.debug("--> Updating task from Google")
                 i_task = i_task.fetch()
                 i_task.synced = datetime.now()
-                i_notion_task: NotionTask = next(NotionTaskRepository.find(notion_id=i_task.notion_id))
+                
+                if sync_notion:
+                    i_notion_task: NotionTask = next(NotionTaskRepository.find(notion_id=i_task.notion_id))
 
-                i_notion_task = i_notion_task.update_from_params(
-                    google_to_notion_task(i_task).dict(
-                        exlude={*NotionTask.Meta.internal_fields}
+                    i_notion_task = i_notion_task.update_from_params(
+                        google_to_notion_task(i_task).dict(
+                            exlude={*NotionTask.Meta.internal_fields}
+                        )
                     )
-                )
-                i_notion_task = i_notion_task.notion_save()
+                    i_notion_task = i_notion_task.notion_save()
+                    i_notion_task = NotionTaskRepository.save(i_notion_task)
+
                 i_task = GoogleTaskRepository.save(i_task)
-                i_notion_task = GoogleTaskRepository.save(i_notion_task)
                 return i_task
 
             elif g_task.updated < i_task.updated:
                 logger.debug("<-- Updating Google task")
-                notion_task: NotionTask = next(NotionTaskRepository.find(
-                    notion_id=i_task.notion_id
-                ))
                 i_task.google_save()
-
+                
                 sync_time = datetime.now()
                 i_task.synced = sync_time
-                notion_task.synced = sync_time
+
+                if sync_notion:
+                    notion_task: NotionTask = next(NotionTaskRepository.find(
+                        notion_id=i_task.notion_id
+                    ))
+                    notion_task.synced = sync_time
+                    notion_task = NotionTaskRepository.save(notion_task)
 
                 i_task = GoogleTaskRepository.save(i_task)
-                notion_task = NotionTaskRepository.save(notion_task)
                 return i_task
             
             else:
@@ -69,12 +74,15 @@ class GoogleSyncer:
 
             try:
                 g_task.synced = datetime.now()
-                notion_task = google_to_notion_task(g_task)
-                notion_task = notion_task.notion_save()
+                
+                if sync_notion:
+                    notion_task = google_to_notion_task(g_task)
+                    notion_task = notion_task.notion_save()
 
-                g_task.notion_id = notion_task.notion_id
+                    g_task.notion_id = notion_task.notion_id
+                    notion_task = NotionTaskRepository.save(notion_task)
+
                 g_task = GoogleTaskRepository.save(g_task)
-                notion_task = NotionTaskRepository.save(notion_task)
                 return g_task
 
             except RuntimeError:
@@ -82,22 +90,29 @@ class GoogleSyncer:
                     # Get parent task from Google and sync parent
                     google_parent_task = GoogleTasks.get(
                         g_task.tasklist, g_task.parent)
-                    synced_parent = self.sync_task(google_parent_task)
+                    synced_parent = self.sync_task(google_parent_task, sync_notion=sync_notion)
                     self.synced_tasks.append(synced_parent)
 
                     # Try again
-                    return self.sync_task(g_task)
+                    return self.sync_task(g_task, sync_notion=sync_notion)
 
                 else:
                     logger.warn("Parent task did not exist in Notion, jumping this one for now")
                     return
 
-    def sync(self):
+    def sync(self, tasklists:List[str]=None, sync_notion=True):
         logger.debug("Syncing tasks FROM Google")
         self.synced_tasks = []
+        google_tasks_to_sync = []
 
-        for g_task in GoogleTasks.list():
-            synced_task = self.sync_task(g_task)
+        if tasklists:
+            for tasklist_id in tasklists:
+                google_tasks_to_sync += list(GoogleTasks.list(tasklist_id=tasklist_id))
+        else:
+            google_tasks_to_sync = list(GoogleTasks.list())
+
+        for g_task in google_tasks_to_sync:
+            synced_task = self.sync_task(g_task, sync_notion=sync_notion)
             self.synced_tasks.append(synced_task)
 
         synced_task_ids = [task.google_id for task in self.synced_tasks]
@@ -108,15 +123,18 @@ class GoogleSyncer:
                 # and external Notion task as well
                 logger.debug('--x ({i_task.title}) Task removed in Google')
                 try:
-                    i_notion_task: NotionTask = next(NotionTaskRepository.find(
-                        notion_id=i_task.notion_id
-                    ))
-                    i_notion_task.notion_delete()
-                    n_deleted = NotionTaskRepository._get_collection().delete_one({"notion_id": i_notion_task.notion_id})
+                    if sync_notion:
+                        i_notion_task: NotionTask = next(NotionTaskRepository.find(
+                            notion_id=i_task.notion_id
+                        ))
+                        i_notion_task.notion_delete()
+                        n_deleted = NotionTaskRepository._get_collection().delete_one({"notion_id": i_notion_task.notion_id})
+
+                        if n_deleted.deleted_count == 0:
+                            logger.error(f'Could not delete the corresponding internal Notion task of "{i_notion_task.title}" (nid={i_notion_task.notion_id})')
+                    
                     g_deleted = GoogleTaskRepository._get_collection().delete_one({"google_id": i_task.google_id})
 
-                    if n_deleted.deleted_count == 0:
-                        logger.error(f'Could not delete the corresponding internal Notion task of "{i_notion_task.title}" (nid={i_notion_task.notion_id})')
                     if g_deleted.deleted_count == 0:
                         logger.error(f'Could not delete the corresponding internal Google task of "{i_task.title}" (gid={i_task.google_id})')
